+++ include/thread.h
@@ -111,6 +111,7 @@ int myabort_checkpoint (hashcat_ctx_t *hashcat_ctx);
 int myabort_finish (hashcat_ctx_t *hashcat_ctx);
 int myabort (hashcat_ctx_t *hashcat_ctx);
 int myquit (hashcat_ctx_t *hashcat_ctx);
+int live_seek (hashcat_ctx_t *hashcat_ctx, const char *request);
 int bypass (hashcat_ctx_t *hashcat_ctx);
 int SuspendThreads (hashcat_ctx_t *hashcat_ctx);
 int ResumeThreads (hashcat_ctx_t *hashcat_ctx);